REFACTOR: 게시글 컨트롤러 비즈니스 로직 걷어내기 4, 빌드

+++ src/main/java/com/example/boardapi/service/CommentService.java
@@ -3,6 +3,7 @@ package com.example.boardapi.service;
 import com.example.boardapi.dto.comment.request.CommentCreateRequestDto;
 import com.example.boardapi.dto.comment.request.CommentEditRequestDto;
 import com.example.boardapi.dto.comment.response.CommentCreateResponseDto;
+import com.example.boardapi.dto.comment.response.CommentEditResponseDto;
 import com.example.boardapi.dto.comment.response.CommentRetrieveResponseDto;
 import com.example.boardapi.entity.Board;
 import com.example.boardapi.entity.Comment;
@@ -19,7 +20,6 @@ import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -37,9 +37,9 @@ public class CommentService {
      * 댓글 저장
      */
     @Transactional
-    public CommentCreateResponseDto save(Long boardId, CommentCreateRequestDto commentCreateRequestDto, HttpServletRequest request) {
+    public CommentCreateResponseDto save(Long boardId, CommentCreateRequestDto commentCreateRequestDto, String token) {
         //글쓴이의 정보(토큰의 정보)
-        Member member = jwtTokenProvider.getMember(request);
+        Member member = jwtTokenProvider.getMember(token);
 
         //게시글 엔티티 조회
         Board board = boardRepository
@@ -119,32 +119,64 @@ public class CommentService {
      * 댓글 수정
      */
     @Transactional
-    public Comment editComment(Long id, CommentEditRequestDto commentEditRequestDto) {
-        Comment comment = retrieveOne(id);
+    public CommentEditResponseDto editComment(Long commentId, Long boardId, CommentEditRequestDto commentEditRequestDto, String token) {
+
+        Member member = jwtTokenProvider.getMember(token);
+
+        //게시글이 존재하는지 검사
+        Board board = boardRepository.findById(boardId).orElseThrow(
+                () -> {
+                    throw new BoardNotFoundException(BoardExceptionMessage.BOARD_NOT_FOUND);
+                }
+        );
+
+        Comment comment = retrieveOne(commentId);
+
+        if (comment.getMember().getId() != member.getId()) {
+            throw new NotOwnBoardException("게시글의 권한이 없습니다.");
+        }
+
         if (comment.isSelected()) {
             throw new InValidUpdateException(CommentExceptionMessage.INVALID_COMMENT_UPDATE);
         }
         comment.changeContent(commentEditRequestDto.getContent());
-        return comment;
+
+        CommentEditResponseDto commentEditResponseDto = modelMapper.map(comment, CommentEditResponseDto.class);
+        commentEditResponseDto.setAuthor(comment.getMember().getName());
+        commentEditResponseDto.setBoardId(boardId);
+        commentEditResponseDto.setMemberId(member.getId());
+        commentEditResponseDto.setSelected(comment.isSelected());
+
+        return commentEditResponseDto;
     }
 
     /**
      * 댓글 삭제
      */
     @Transactional
-    public void deleteComment(Board board, Long id) {
+    public void deleteComment(Long boardId, Long id, String token) {
+
+        //게시글이 존재하는지 검사
+        Board board = boardRepository.findById(boardId).orElseThrow(
+                () -> {
+                    throw new BoardNotFoundException(BoardExceptionMessage.BOARD_NOT_FOUND);
+                }
+        );
+
+        //댓글이 존재하는 확인
         Comment comment = retrieveOne(id);
+
+        Member member = jwtTokenProvider.getMember(token);
+
+        if (comment.getMember().getId() != member.getId()) {
+            throw new NotOwnBoardException("게시글의 권한이 없습니다.");
+        }
 
         if (comment.isSelected()) {
             throw new InValidUpdateException(CommentExceptionMessage.INVALID_COMMENT_UPDATE);
         }
 
-        try {
-            commentRepository.deleteById(id);
-
-        } catch (IllegalArgumentException e) {
-            throw new CommentNotFoundException(CommentExceptionMessage.COMMENT_NOT_FOUND);
-        }
+        commentRepository.deleteById(id);
 
         board.decreaseComments();
     }
@@ -154,10 +186,28 @@ public class CommentService {
     }
 
     @Transactional
-    public void updateCommentLike(Member member, Long commentId) {
+    public void updateCommentLike(Long boardId, Long commentId, String token) {
+        Member member = jwtTokenProvider.getMember(token);
+
+        //게시글이 존재하는지 검사
+        Board board = boardRepository.findById(boardId).orElseThrow(
+                () -> {
+                    throw new BoardNotFoundException(BoardExceptionMessage.BOARD_NOT_FOUND);
+                }
+        );
+
+        if (member.getLikeId().contains(commentId)) {
+            throw new DuplicatedLikeException("이미 좋아요를 눌렀습니다.");
+        }
+
+        //회원의 좋아요 id 추가
         member.getLikeId().add(commentId);
+        
+        //댓글 엔티티 조회
         Comment comment = retrieveOne(commentId);
         int like = comment.getLikes();
+        
+        //댓글 좋아요 수 1 증가
         comment.setLikes(++like);
     }
 
@@ -167,9 +217,9 @@ public class CommentService {
     }
 
     @Transactional
-    public void selectComment(Long boardId, Long commentId, HttpServletRequest request) {
+    public void selectComment(Long boardId, Long commentId, String token) {
         //request 객체의 헤더 부분에서 회원 조회
-        Member member = jwtTokenProvider.getMember(request);
+        Member member = jwtTokenProvider.getMember(token);
 
         //해당 게시글에도 채택 됨을 알기 위해 조회
         Board board = boardRepository.findById(boardId).orElseThrow(
